Fix superfluous warning due to race in gunzip
Since reading the header closes the pipe before it is done being used,
this can result in the gunzip thread failing with an exception about the
closed port.  Ideally, we would catch only this error, but it raises
exn:fail.  However, the error printed is misleading, so catch it and
avoid printing it.